# commons/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "IOTGlue NodeAPI's",
    "description": "",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/iotglueapi/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/rtbuilder/rules": {
      "post": {
        "description": "",
        "summary": "createRule",
        "tags": [
          "Rule"
        ],
        "operationId": "createRule",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createRuleRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "",
        "summary": "getAllRulesWithPagination",
        "tags": [
          "Rule"
        ],
        "operationId": "getAllRulesWithPagination",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page_no",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "page_size",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }},
    "/rtbuilder/rules/{ruleId}": {
      "get": {
        "description": "",
        "summary": "getRuleById",
        "tags": [
          "Rule"
        ],
        "operationId": "getRuleById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "ruleId",
            "in": "path",
            "description": "ID of rule to return",
            "required": true,
            "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "",
        "summary": "updateRuleById",
        "tags": [
          "Rule"
        ],
        "operationId": "updateRuleById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "ruleId",
            "in": "path",
            "description": "ID of rule to update",
            "required": true,
            "type": "string"
            },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateRuleByIdRequest"
            }
          }
          
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "",
        "summary": "deleteRuleById",
        "tags": [
          "Rule"
        ],
        "operationId": "deleteRuleById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "ruleId",
          "in": "path",
          "description": "ID of rule to delete",
          "required": true,
          "type": "string"
          }],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/common/rightpanel/rules/{ruleId}": {
      "get": {
        "description": "getSourceForRightPanelById",
        "summary": "getRuleForRightPanelById",
        "tags": [
          "Common - Right Panel"
        ],
        "operationId": "getRuleForRightPanelById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [{
            "name": "ruleId",
            "in": "path",
            "description": "ID of Rule",
            "required": true,
            "type": "string"
            }],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/common/rules/filter": {
      "post": {
        "summary": "getFilteredRules",
        "tags": [
          "Common"
        ],
        "operationId": "getFilteredRules",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/getFilteredRulesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
},
  "definitions": {
...
    "createRuleRequest": {
      "title": "createRuleRequest",
      "example": {
        "description": "script test",
        "nodeType": "Rule",
        "sourceType": "Device",
        "sourceId": "C_DR_5",
        "type": "script",
        "protocol": "MQTT",
        "simpleExp": "",
        "complexExp": "",
        "script": "RTB_SC_6",
        "status": "Approved",
        "topicOrVersionChosenOrUri": "Fleet Topic-1",
        "createdBy": "ajayn",
        "createdOn": "2019-09-30T09:50:27.403Z",
        "modifiedOn": "",
        "modifiedBy": "",
        "name": "OLA script",
        "statusUpdateComment": ""
      },
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "nodeType": {
          "type": "string"
        },
        "sourceType": {
          "type": "string"
        },
        "sourceId": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "protocol": {
          "type": "string"
        },
        "simpleExp": {
          "type": "string"
        },
        "complexExp": {
          "type": "string"
        },
        "script": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "topicOrVersionChosenOrUri": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "createdOn": {
          "type": "string"
        },
        "modifiedOn": {
          "type": "string"
        },
        "modifiedBy": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "statusUpdateComment": {
          "type": "string"
        }
      },
      "required": [
        "description",
        "nodeType",
        "sourceType",
        "sourceId",
        "type",
        "protocol",
        "simpleExp",
        "complexExp",
        "script",
        "status",
        "topicOrVersionChosenOrUri",
        "createdBy",
        "createdOn",
        "modifiedOn",
        "modifiedBy",
        "name",
        "statusUpdateComment"
      ]
    },
    "updateRuleByIdRequest": {
      "title": "updateRuleByIdRequest",
      "example": {
        "description": "script test",
        "nodeType": "Rule",
        "sourceType": "Device",
        "sourceId": "C_DR_5",
        "type": "script",
        "protocol": "MQTT",
        "simpleExp": "",
        "complexExp": "",
        "script": "RTB_SC_6",
        "status": "Approved",
        "topicOrVersionChosenOrUri": "Fleet Topic-1",
        "createdBy": "ajayn",
        "createdOn": "2019-09-30T09:50:27.403Z",
        "modifiedOn": "",
        "modifiedBy": "",
        "name": "OLA script",
        "statusUpdateComment": ""
      },
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "nodeType": {
          "type": "string"
        },
        "sourceType": {
          "type": "string"
        },
        "sourceId": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "protocol": {
          "type": "string"
        },
        "simpleExp": {
          "type": "string"
        },
        "complexExp": {
          "type": "string"
        },
        "script": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "topicOrVersionChosenOrUri": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "createdOn": {
          "type": "string"
        },
        "modifiedOn": {
          "type": "string"
        },
        "modifiedBy": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "statusUpdateComment": {
          "type": "string"
        }
      },
      "required": [
        "description",
        "nodeType",
        "sourceType",
        "sourceId",
        "type",
        "protocol",
        "simpleExp",
        "complexExp",
        "script",
        "status",
        "topicOrVersionChosenOrUri",
        "createdBy",
        "createdOn",
        "modifiedOn",
        "modifiedBy",
        "name",
        "statusUpdateComment"
      ]
    },
...
    "getFilteredRulesRequest": {
      "title": "getFilteredRulesRequest",
      "example": {
        "sourceId": "SG_1",
        "status": "Approved",
        "topicOrUri": "/data/1",
        "templateBased": true
      },
      "type": "object",
      "properties": {
        "sourceId": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "topicOrUri": {
          "type": "string"
        },
        "templateBased": {
          "type": "boolean"
        }
      },
      "required": [
        "sourceId",
        "status",
        "topicOrUri",
        "templateBased"
      ]
    }
  },
...
}